Separation of concerns, and pending tests
