+++ app/styles/score-box.css
@@ -1,11 +1,14 @@
 #score-box {
+  border-width: 1px;
+  border-radius: 10px;
+  padding: 2px 5px 2px 5px;
+}
+
+#score-box.invalid {
+  border-color: black;
   background-color: white;
   color: black;
   border: solid;
-  border-width: 1px;
-  border-radius: 10px;
-  border-color: black;
-  padding: 2px 5px 2px 5px;
 }
 
 #score-box.high {
